re-use `getPossibleEventMessages` util
