+++ packages/core/src/utils/eventUtils.ts
@@ -1,0 +1,31 @@
+import type { Event } from '../types-hoist';
+
+/**
+ * Get a list of possible event messages from a Sentry event.
+ */
+export function getPossibleEventMessages(event: Event): string[] {
+  const possibleMessages: string[] = [];
+
+  if (event.message) {
+    possibleMessages.push(event.message);
+  }
+
+  let lastException;
+  try {
+    // @ts-expect-error Try catching to save bundle size
+    lastException = event.exception.values[event.exception.values.length - 1];
+  } catch (e) {
+    // try catching to save bundle size checking existence of variables
+  }
+
+  if (lastException) {
+    if (lastException.value) {
+      possibleMessages.push(lastException.value);
+      if (lastException.type) {
+        possibleMessages.push(`${lastException.type}: ${lastException.value}`);
+      }
+    }
+  }
+
+  return possibleMessages;
+}
